+++ app/Models/Resume/Project.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace App\Models;
+namespace App\Models\Resume;
 
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Company extends Model
+class Project extends Model
 {
     use HasFactory;
     use HasUuids;
@@ -16,11 +16,8 @@ class Company extends Model
     protected $guarded = ['id'];
     public $timestamps = false;
 
-    /**
-     * Get the projects for the company.
-     */
-    public function projects()
+    public function company()
     {
-        return $this->hasMany(Project::class);
+        return $this->belongsTo(Company::class);
     }
 }
